requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
numpy>=1.26.0
